# Update to support opentracing.SpanContext (and References)