--- a/concurrency_test.go
+++ b/concurrency_test.go
@@ -83,11 +83,11 @@ func TestConcurrentUsage(t *testing.T) {
 				sp := tracer.StartSpan(op)
 				sp.LogEvent("test event")
 				sp.SetTag("foo", "bar")
-				sp.SetBaggageItem("boo", "far")
+				sp.Context().SetBaggageItem("boo", "far")
 				sp.SetOperationName("x")
-				csp := tracer.StartSpanWithOptions(opentracing.StartSpanOptions{
-					Parent: sp,
-				})
+				csp := tracer.StartSpan(
+					"csp",
+					opentracing.ChildOf(sp.Context()))
 				csp.Finish()
 				defer sp.Finish()
 			}
@@ -105,9 +105,8 @@ func TestDisableSpanPool(t *testing.T) {
 	parent := tracer.StartSpan("parent")
 	parent.Finish()
 	// This shouldn't panic.
-	child := tracer.StartSpanWithOptions(opentracing.StartSpanOptions{
-		Parent:        parent,
-		OperationName: "child",
-	})
+	child := tracer.StartSpan(
+		"child",
+		opentracing.ChildOf(parent.Context()))
 	child.Finish()
 }