--- a/propagation_test.go
+++ b/propagation_test.go
@@ -13,7 +13,7 @@ import (
 )
 
 type verbatimCarrier struct {
-	basictracer.Context
+	basictracer.SpanContext
 	b map[string]string
 }
 
@@ -30,11 +30,11 @@ func (vc *verbatimCarrier) GetBaggage(f func(string, string)) {
 }
 
 func (vc *verbatimCarrier) SetState(tID, sID uint64, sampled bool) {
-	vc.Context = basictracer.Context{TraceID: tID, SpanID: sID, Sampled: sampled}
+	vc.SpanContext = basictracer.SpanContext{TraceID: tID, SpanID: sID, Sampled: sampled}
 }
 
 func (vc *verbatimCarrier) State() (traceID, spanID uint64, sampled bool) {
-	return vc.Context.TraceID, vc.Context.SpanID, vc.Context.Sampled
+	return vc.SpanContext.TraceID, vc.SpanContext.SpanID, vc.SpanContext.Sampled
 }
 
 func TestSpanPropagator(t *testing.T) {
@@ -43,7 +43,7 @@ func TestSpanPropagator(t *testing.T) {
 	tracer := basictracer.New(recorder)
 
 	sp := tracer.StartSpan(op)
-	sp.SetBaggageItem("foo", "bar")
+	sp.Context().SetBaggageItem("foo", "bar")
 
 	tmc := opentracing.HTTPHeaderTextMapCarrier(http.Header{})
 	tests := []struct {
@@ -55,13 +55,16 @@ func TestSpanPropagator(t *testing.T) {
 	}
 
 	for i, test := range tests {
-		if err := tracer.Inject(sp, test.typ, test.carrier); err != nil {
+		if err := tracer.Inject(sp.Context(), test.typ, test.carrier); err != nil {
 			t.Fatalf("%d: %v", i, err)
 		}
-		child, err := tracer.Join(op, test.typ, test.carrier)
+		injectedContext, err := tracer.Extract(test.typ, test.carrier)
 		if err != nil {
 			t.Fatalf("%d: %v", i, err)
 		}
+		child := tracer.StartSpan(
+			op,
+			opentracing.ChildOf(injectedContext))
 		child.Finish()
 	}
 	sp.Finish()