--- a/raw.go
+++ b/raw.go
@@ -8,9 +8,13 @@ import (
 
 // RawSpan encapsulates all state associated with a (finished) Span.
 type RawSpan struct {
-	// The RawSpan embeds its Context. Those recording the RawSpan
-	// should also record the contents of its Context.
-	Context
+	// The RawSpan embeds its SpanContext. Those recording the RawSpan
+	// should also record the contents of its SpanContext.
+	SpanContext
+
+	// The SpanID of this SpanContext's first intra-trace reference (i.e.,
+	// "parent"), or 0 if there is no parent.
+	ParentSpanID uint64
 
 	// The name of the "operation" this span is an instance of. (Called a "span
 	// name" in some implementations)
@@ -27,7 +31,4 @@ type RawSpan struct {
 
 	// The span's "microlog".
 	Logs []opentracing.LogData
-
-	// The span's associated baggage.
-	Baggage map[string]string // initialized on first use
 }