--- a/raw.go
+++ b/raw.go
@@ -10,7 +10,7 @@ import (
 type RawSpan struct {
 	// The RawSpan embeds its SpanContext. Those recording the RawSpan
 	// should also record the contents of its SpanContext.
-	SpanContext
+	*SpanContext
 
 	// The SpanID of this SpanContext's first intra-trace reference (i.e.,
 	// "parent"), or 0 if there is no parent.