--- a/recorder_test.go
+++ b/recorder_test.go
@@ -12,7 +12,7 @@ func TestInMemoryRecorderSpans(t *testing.T) {
 	recorder := NewInMemoryRecorder()
 	var apiRecorder SpanRecorder = recorder
 	span := RawSpan{
-		SpanContext: SpanContext{},
+		SpanContext: &SpanContext{},
 		Operation:   "test-span",
 		Start:       time.Now(),
 		Duration:    -1,