--- a/tracer.go
+++ b/tracer.go
@@ -118,11 +118,13 @@ type tracerImpl struct {
 
 func (t *tracerImpl) StartSpan(
 	operationName string,
+	opts ...opentracing.StartSpanOption,
 ) opentracing.Span {
-	return t.StartSpanWithOptions(
-		opentracing.StartSpanOptions{
-			OperationName: operationName,
-		})
+	sso := opentracing.StartSpanOptions{}
+	for _, o := range opts {
+		o.Apply(&sso)
+	}
+	return t.StartSpanWithOptions(operationName, sso)
 }
 
 func (t *tracerImpl) getSpan() *spanImpl {
@@ -135,6 +137,7 @@ func (t *tracerImpl) getSpan() *spanImpl {
 }
 
 func (t *tracerImpl) StartSpanWithOptions(
+	operationName string,
 	opts opentracing.StartSpanOptions,
 ) opentracing.Span {
 	// Start time.
@@ -147,31 +150,45 @@ func (t *tracerImpl) StartSpanWithOptions(
 	tags := opts.Tags
 
 	// Build the new span. This is the only allocation: We'll return this as
-	// a opentracing.Span.
+	// an opentracing.Span.
 	sp := t.getSpan()
-	if opts.Parent == nil {
+	// Look for a parent in the list of References.
+	//
+	// TODO: would be nice if basictracer did something with all
+	// References, not just the first one.
+ReferencesLoop:
+	for _, ref := range opts.References {
+		switch ref.Type {
+		case opentracing.ChildOfRef,
+			opentracing.FollowsFromRef:
+
+			refMD := ref.Referee.(*SpanContext)
+			sp.raw.TraceID = refMD.TraceID
+			sp.raw.SpanID = randomID()
+			sp.raw.ParentSpanID = refMD.SpanID
+			sp.raw.Sampled = refMD.Sampled
+
+			refMD.baggageLock.Lock()
+			if l := len(refMD.Baggage); l > 0 {
+				sp.raw.Baggage = make(map[string]string, len(refMD.Baggage))
+				for k, v := range refMD.Baggage {
+					sp.raw.Baggage[k] = v
+				}
+			}
+			refMD.baggageLock.Unlock()
+			break ReferencesLoop
+		}
+	}
+	if sp.raw.TraceID == 0 {
+		// No parent Span found; allocate new trace and span ids and determine
+		// the Sampled status.
 		sp.raw.TraceID, sp.raw.SpanID = randomID2()
 		sp.raw.Sampled = t.options.ShouldSample(sp.raw.TraceID)
-	} else {
-		pr := opts.Parent.(*spanImpl)
-		sp.raw.TraceID = pr.raw.TraceID
-		sp.raw.SpanID = randomID()
-		sp.raw.ParentSpanID = pr.raw.SpanID
-		sp.raw.Sampled = pr.raw.Sampled
-
-		pr.Lock()
-		if l := len(pr.raw.Baggage); l > 0 {
-			sp.raw.Baggage = make(map[string]string, len(pr.raw.Baggage))
-			for k, v := range pr.raw.Baggage {
-				sp.raw.Baggage[k] = v
-			}
-		}
-		pr.Unlock()
 	}
 
 	return t.startSpanInternal(
 		sp,
-		opts.OperationName,
+		operationName,
 		startTime,
 		tags,
 	)
@@ -203,28 +220,28 @@ type delegatorType struct{}
 // Delegator is the format to use for DelegatingCarrier.
 var Delegator delegatorType
 
-func (t *tracerImpl) Inject(sp opentracing.Span, format interface{}, carrier interface{}) error {
+func (t *tracerImpl) Inject(sc opentracing.SpanContext, format interface{}, carrier interface{}) error {
 	switch format {
 	case opentracing.TextMap:
-		return t.textPropagator.Inject(sp, carrier)
+		return t.textPropagator.Inject(sc, carrier)
 	case opentracing.Binary:
-		return t.binaryPropagator.Inject(sp, carrier)
+		return t.binaryPropagator.Inject(sc, carrier)
 	}
 	if _, ok := format.(delegatorType); ok {
-		return t.accessorPropagator.Inject(sp, carrier)
+		return t.accessorPropagator.Inject(sc, carrier)
 	}
 	return opentracing.ErrUnsupportedFormat
 }
 
-func (t *tracerImpl) Join(operationName string, format interface{}, carrier interface{}) (opentracing.Span, error) {
+func (t *tracerImpl) Extract(format interface{}, carrier interface{}) (opentracing.SpanContext, error) {
 	switch format {
 	case opentracing.TextMap:
-		return t.textPropagator.Join(operationName, carrier)
+		return t.textPropagator.Extract(carrier)
 	case opentracing.Binary:
-		return t.binaryPropagator.Join(operationName, carrier)
+		return t.binaryPropagator.Extract(carrier)
 	}
 	if _, ok := format.(delegatorType); ok {
-		return t.accessorPropagator.Join(operationName, carrier)
+		return t.accessorPropagator.Extract(carrier)
 	}
 	return nil, opentracing.ErrUnsupportedFormat
 }