See https://github.com/opentracing/opentracing-go/pull/82 .

This change is included to show some of the ramifications of `SpanContext` for implementations and OT callers (via the examples and tests here). I'll add a few notes inline.

Benchmark diffs:

```
bhs@bhs-personal:~/git/lightstep/go/src$ ../bin/benchcmp /tmp/before /tmp/after
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkSpan_Empty-8                                        629           632           +0.48%
BenchmarkSpan_100Events-8                                    35847         35271         -1.61%
BenchmarkSpan_1000Events-8                                   35587         35173         -1.16%
BenchmarkSpan_100Tags-8                                      43618         44145         +1.21%
BenchmarkSpan_1000Tags-8                                     44063         43838         -0.51%
BenchmarkSpan_100BaggageItems-8                              145397        32579         -77.59%
BenchmarkTrimmedSpan_100Events_100Tags_100BaggageItems-8     168957        79144         -53.16%
BenchmarkInject_TextMap_Empty-8                              2221          2292          +3.20%
BenchmarkInject_TextMap_100BaggageItems-8                    85977         87857         +2.19%
BenchmarkInject_Binary_Empty-8                               388           390           +0.52%
BenchmarkInject_Binary_100BaggageItems-8                     17861         18968         +6.20%
BenchmarkJoin_TextMap_Empty-8                                2849          2872          +0.81%
BenchmarkJoin_TextMap_100BaggageItems-8                      148235        146900        -0.90%
BenchmarkJoin_Binary_Empty-8                                 1175          1231          +4.77%
BenchmarkJoin_Binary_100BaggageItems-8                       37424         37700         +0.74%
```

I made these changes pretty casually and wonder if the huge improvements on the BaggageItems benchmarks are due to some sort of measurement error in the test itself (i.e., an invalid result).