@bensigelman 2016-06-15
... this PR has also been updated per https://github.com/opentracing/opentracing-go/pull/82.