@bensigelman 2016-07-03
This has been updated per https://github.com/opentracing/opentracing.github.io/pull/100 and #32