@bensigelman 2016-07-04
I will merge this (along with https://github.com/opentracing/opentracing-go/pull/82) tomorrow if there are no objections.