@bensigelman 2016-07-06
Well, `go vet` did a good job finding some actual problems here... namely that it's bad to copy structs by value that have `Mutex`es inside.

The `Mutex` was there to ensure that the `Baggage` accesses are concurrency-safe. Since the `SpanContext` may now outlive the `Span`, the easy+safe fix is to make it a pointer within `Span`. Sadly this does mean that there will be more allocations.

We could create a pool for these (and that would make sense), but for now I would rather get this change in and unbreak builds that depend on `basictracer-go` and `opentracing-go`.